Create and use a password entry component
